requirements.txt
certifi            2024.12.14
charset-normalizer 3.4.1
colorama           0.4.6
contourpy          1.3.1
cycler             0.12.1
filelock           3.13.1
fonttools          4.55.3
fsspec             2024.2.0
gitdb              4.0.12
GitPython          3.1.44
idna               3.10
Jinja2             3.1.3
kiwisolver         1.4.8
MarkupSafe         2.1.5
matplotlib         3.10.0
mpmath             1.3.0
networkx           3.2.1
numpy              1.26.3
opencv-python      4.10.0.84
packaging          24.2
pandas             2.2.3
pillow             11.1.0
pip                24.0
psutil             6.1.1
py-cpuinfo         9.0.0
pyparsing          3.2.1
python-dateutil    2.9.0.post0
pytz               2024.2
PyYAML             6.0.2
requests           2.32.3
scipy              1.15.0
seaborn            0.13.2
setuptools         75.8.0
six                1.17.0
smmap              5.0.2
sympy              1.13.1
thop               0.1.1.post2209072238
torch              2.5.1+cu124
torchaudio         2.5.1+cu124
torchvision        0.20.1+cu124
tqdm               4.67.1
typing_extensions  4.9.0
tzdata             2024.2
ultralytics        8.3.59
ultralytics-thop   2.0.13
urllib3            2.3.0